Fixing a rare failure mode. This failure mode was exhibited by iqtree -s example.phy -m LM5.7b -redo -seed 1 --eigenlib (by the nature of failure mode, the input file should not matter.)
The problem is that the rate matrix can be non-diagonalizable, at which
point numeric noise takes over and can wreack havoc on your matrix
exponentiation, giving a non-stochastic transition matrix.

The solution is to check for non-diagonalizability (signaled by the
eigenvector matrix having very small determinant) and use the scaled squaring
matrix exponentiation method as a fallback should this occur.