Cleaned code; some testing; heterozygosity.
There is a difficulty when calculating the heterozygosity of the
boundary mutation model.  Only when N is large, the heterozygosity is
close to the empricial one.  This needs to be looked at.  At N=3, the
heterozygosity is off by a factor of 2.  In general, this is not a
problem; but when the heterozygosity is fixed (either because the
value is known or estimated from the data), it leads to wrong
estimations.