Bugfix, interpretation of +P.
Minh, please have a short look at the TODO Minh items. The +P flag was not found
when it was at the first position of the string. This led to errors, which are
fixed now. However, you might have had a reason to start search for +P only at
the second position in the given model name.