Corrected the mistake that was making BIONJ-R and NJ-R tree construction go wrong: In BoundingMatrix's getRowMinima member function, maxEarlierTotal, which has to be declared an NJFloat, was declared as size_t (so it was getting set to zero). The net effect was that the first entry looked at in the S matrix always "won" (and was treated as though it corresponded to the minimum entry in the Q matrix).
Corrected two vector initialisations that assumed the width of the V/FloatVector
template parameter was always 4.  Since, when you're using T=single, V=Vec8f, the
Width of V is 8.  In getRowMinima member function, in VectorizedMatrix.
I was already using the scalar assignment operator in VectorizedUPGMA_Matrix,
which doesn't care about the width of V.

Totalling of rows of the D matrix (which happens after clustering,
For the row for the new cluster in UPGMA, NJ, and BIONJ), moved to Matrix.
This should help performance for UPGMA_Matrix and NJMatrix, a little,
on large inputs
(It won't help BIONJ, since that was the "donor" class for the code).

removeRowAndColumn now does row copying rather than pointer reassignment.
removeRowAndColumn also moves rows closer together (grouping them all
close to the front of the allocated array) periodically.
Both of these changes reduce the amount of memory in use and should
help performance on very large inputs. On my laptop, the break-even
point is somewhere between 5K and 20K rows.