Minor tuneups in Alignment::checkSeqName (vectorisation) and Alignment::buildPattern (parallelization)... 1. Added sumForUnknownCharacters (in hammingdistance.h)     (for a given sequence, it sums up the frequencies for the unknown characters, using Vec32c character vectors). 2. Added Pattern::isAllGaps (in pattern.h and pattern.cpp)     (it determines whether a pattern consists entirely of gap (unknown) states) 3. Alignment::checkSeqName now makes use of sumForUnknownCharacters     rather than doing the frequency summation for unknown characters itself. 4. Alignment::buildPattern now does more of its processing in parallel.     (calculating patterns for each site, and then sequentially identifying     patterns that can be "compressed").  This change is mostly aimed at     improving performance in debug builds on windows (where it reduces the     running time for Alignment::buildPattern by about 25%, on a 4-core windows     laptop) (it doesn't matter much for release builds or on other Operating Systems).
Note that Alignment::buildPattern could be tuned further (hashing is being done in
the sequential part of the code, and the hashing could be done *outside* of the
hash map beng used for the compression), but I haven't rigged that yet.  Furthermore,
Alignment::extractSubAlignment has similar problems (it is still doing most of its
processing "sequentially", because I haven't made changes to it like those in change #4).
When I get to that, I'll probably move the pattern compression logic from buildPattern
into a separate compressPatterns member function, which can be called from both
buildPattern and extractSubAlignment.

There *is* a trade-off here. Pattern compression in serial "as they are processed"
takes longer, but uses (a bit) less memory than "load the patterns in parallel, and
then compress them".