Fix for defect introduced in commit ef7636a: computeConst wasn't ever called for patterns being examined in Alignment::buildPattern.
(It's being called for every pattern, before compression - before patterns that
are duplicates of earlier patterns are removed.  It could also
be called only for patterns we've decided to keep, next to the at(w).frequency = 1
line lower down. I've tried both, and calling it for every pattern, before
the identification of duplicates, is quite a bit faster) (except when the thread
count is 1, in which case it is just as fast)

(I also tried doing it in a third stage, after compression, in another parallel for,
Only on the kept patterns, and that was slower yet!  Worse temporal locality
of memory references!)