Drafting support for incremental processing (updating of) user-defined trees and checkpoint trees.  The new code *only* executes if -incremental parameter is passed on the command-line.  This code is *not* functional as yet. It crashes (if -incremental is passed and the input tree's Sequence names don't match those of the alignment): in reorientPartialLh (I haven't figured out why).
The important new functions are
PhyloTree::updateToMatchAlignment, which removes taxa
(In the tree) that no longer appear in the alignment, and
then calls on PhyloTree::addNewTaxaToTree.

For now, both functions are inflexible: no tidy-up or
Restructuring (even local, like say, local NNI),
after removal of taxa - either local after batches of removals,
or global after all are removed.
And, when adding new taxa: no support for heuristics to
Find points of insertion, no parallelisation (global brute
Force search only!), no choice of local cost function)
(it's hardcoded to be ML, for now at least),
No tidy up (by, say, local NNI?)  after batches of inserts,
No tidy up (or restructuring) after all the inserts are done.

PhyloTree::addNewTaxaToTree is broken.  When it has work to
do, it segfaults in reorientPartialLh (and I have so far been
unable to figure out why).

There's a temporary testing hack in
PhyloTree::updateToMatchAlignment.  If you #define
BODGE_ONE_NODE_IN_TEN to be ... 1, the tree that is read in
has the names of every tenth sequence, starting
with sequence 0 mangled (by appending "_Bodged" to it).
This makes it a lot easier for me to test, for now.

1. Added methods to PhyloTree:
   (a) updateToMatchAlignment (which is the incremental
       Equivalent of setAlignment).
   (b) configureLikelihoodKernel (sets likelihood kernel based
       on command-line parameters)
   (c) configureModel (sets up model factory, model, rate,
       based on command-line parameters)
   (d) initializeModel (this is so that the code in
       updateToMatchAlignment, defined in PhyloTree, can
       call it).
   (e) addNewTaxaToTree
   (f) computeBranchLengths (this is so that the code in
       updateToMatchAlignment - defined in PhyloTree - can
       potentially "call on" code in subclasses that do branch
       length recalculation).
2. removeTaxa (in MTree, and ConstraintTree)
   now takes two additional parameters; one indicating whether
   to reassign id numbers for nodes, to match surviving taxa
   (which you do NOT want to do from under 1(a)), and a
   context parameter (used for reporting progress), which is
   usually nullptr (but not when called from 1(a)).
3. Added hideProgress() progress-displaying definitions etc.
   to MTree (with do-nothing implementations). Those in
   PhyloTree override them.
4. MTree::removeTaxa now reports progress (if its context
   parameter isn't nullptr, via the virtual methods added in
   change 3 above).
5. In IQTree::computeInitialTree(), if Params::incremental is
   set, the code for reading user-supplied input trees calls on
   updateToMatchAlignment (change 1), rather than setAlignment.
   Likewise, the code for restoring a previous tree from a
   checkpoint, calls updateToMatchAlignment too.
6. Added clearAllScaleNum member functions to PhyloNode,
   PhyloTree, and PhyloSuperTree.
7. main/phyloanalysis.cpp - corrected typo in comment,
   tree/phylosupertreeunlinked.h, fixed up an unclear comment.