1. Added a count_unknown_as_different member to Params    (And code to let the user turn it on, via a -cud     command-line parameter) (Not documented yet). 2. count_unknown_as_different is honoured in    AlignmentPairwise::recomputeDist(). 3. It is also honoured in computeDistanceMatrix    (in tree/phylotree.cpp). 4. PhyloTree::addNewTaxa doesn't do its own allocation of partial    likelihood or scale_num vectors any more (since that didn't seem    to help!) (This function only called when -incremental command-line    parameter is passed; and it still doesn't work!). 5. MTree::removeTaxa member function (also its override in    ConstraintTree) now takes a const StrVector, rather than    a StrVector, as its first Parameter). 6. MTree::findLeafName member function    Now takes a const string rather than just a string, as its first    Parameter).
Counting unknown sites as different (from sites that are known
or other sites that are unknown) via -cud seems to help UPGMA get better
answers (sometimes, at least).  It doesn't help NJ or BIONJ at all
(with it, they find trees with lower log-likelihoods).
It (passing the -cud command-line parameter) also helps a
distance matrix algorithm (STITCH), that I have been working on
in my spare time, find trees with higher log-likelihoods.
(STITCH is not *yet* competitive with NJ or BIONJ, and I haven't
committed any of the changes for it to the dev branch yet).