Bug fix in UPGMA_Matrix::getMinimumEntry(), to prevent it from choosing "joining a cluster with itself" as the next join to do, during neighbour joining.
Depending what was on the stack (!) that was sometimes happening,
because a dummy Position had row=0, column=0 (row and column
indicate row and column numbers in a matrix of intra-cluster
distances).  Joining a cluster with itself resulted in that
cluster (and anything in it) not appearing in the output tree at
all (like this: replace A with A2 (joining A with itself), then
remove A2 from consideration).
What should happen is replace A with A2 (joining A with B),
then remove B from consideration).

This affected Rapid NJ, the default tree construction algorithm
used on the distance matrix, as it is based on a subclass of UPGMA_Matrix.