Performance tune-ups to parsimony optimisation and parsimony placement.
1. TargetBranch::computeState now takes parsimony_score
   as a double& (so, if it is supplied -1.0, it can
   tell the caller what the parsimony score really is,
   and if the caller calls it again... it won't have to
   pass -1.0).
2. Various clients of TargetBranch::ComputeState now
   take advantage of change (1):
   (a) PlacementRun::doBatchPlacementCosting()
   (b) BaseballSearchHeuristic::prepareToFilter()
   (c) TargetBranchRange::addNewRef()
   (d) TaxonToPlace::assessNewTargetBranches()
   (e) ParsimonyLazyTBRMove::apply()
   This speeds them all up (a little teensy bit).
3. TargetBranch now has a parsimony_dirtiness member,
   which it can use to "cheat" when calculating
   connection costs (which it only needs to do in
   TargetBranch::computeState when parsimony_score is -1.0).
   This speeds up parsimony placement considerably (x3).
4. TaxonToPlace::assessNewTargetBranches now
   does its assessPlacementCost calls in parallel
   (I hadn't bothered having it do this before, because
   the # of taxa being added to the tree was, until
   recently, always relatively low, so it didn't
   matter that this was sequential).  But if the # of
   taxa to be added is very large (hundreds of times
   the # already in the tree)...
   This sped up parsimony placement by about 10%.
5. PhyloTree::computeParsimonyTreeBatch now uses
   sample trees of size proportional to the square root
   of the size the overall tree.

The main thing slowing down parsimony placement is,
for the subtrees being constructed "late", the time it
takes to calculate the consequences of the additions of
the earlier subtrees increases as placement proceeds.
A parameter to... turn that off (!) (and to "pretend"
that each subtree is the first), would probably not
make the parsimony score much worse, but would avoid
doing those parsimony calculations that take account of
the other trees.  It's only those calculations that
keep the exponent (of N in the running time) above 1.5
Without them, it could be lowered to 1.5.
(And an exponent lower even than that is probably
possible!) (recursion, sampling the items going to a
particular target branch, distributing the others
using the sample).

Once PhyloTree::optimizePlacementRegion is working
properly, that option will be workable.