Draft code for similarity matching trees (which might... someday... lead to an alternative to parsimony and distance matrix tree construction algorithms).
A similarity tree is a rooted binary tree; the leaf nodes
are taxa, the interior nodes are "population averages" of
the nodes in the subtrees below.

The question asked during (similarity) tree
construction is, "which of two subtrees" contains taxa that
are, on average, most like the taxon being added to the tree.
The similarity tree is weight balanced, to ensure an
O(p*n.ln(n)) worst case running time, where p is the number
of parsimony-informative sights, and n is the number of sequences
in the alignment.

For now, the trees aren't much good at all (they are barely
better than random trees) (but it is early days!).

1. Added STT_SIMILARITY_MATCHING to START_TREE_TYPE.
2. Added code in parseTreeName() (in utils/tools.cpp) to treat
   -t SIMA as a request for a similarity matching tree.
3. Added PhyloTree member functions (definitions in tree/phylotree.h,
   implementations in tree/similaritymatchingtree.cpp, for now):
   (a) createSimilarityMatchingTree() (top)
   (b) calculateSimilarityWeights()
   (c) constructSimilarityTree()
   (d) mirrorSimilarityTreeStructure() converts node structure
   (e) correctSimilarityTreeStructure() converts the top of a
       A                rooted binary tree into an unrooted
       |\               tree, at the end of the proess,
       B C   to   A-C   by replacing the top node (and its
       |\         |\    largest child) with a single degree-3
       D E        D E   node.
4.Added code in IQTree::computeInitialTree to call 3(a) when
  the start tree is 1.  Likewise, IQTree::initCandidateTreeSet().
5.The implementations of 3(a) through 3(e) make use of
  WBTNode, WeightBalancedTree template classes, and
  SimilarityNode, all defined in tree/similaritymatchingtree.cpp.
6. Updated some loops (over sequences) in tree/phylotree.cpp to
   work with intptr_t rather than int (or size_t).
7. Likewise, in tree/iqtree.cpp
8. Added some code to IQTree::doTreeSearch() to track how much
   time is being spent on tree perturbation.