Rationalizing Params members.  Changes in header file (utils/tools.h) first... (1 through 4) 1. Made the constructor private (so only getInstance can give    you a Params isntance), and moved its implementation to    utils/tools.cpp 2. Deleted the assignment operator and the copy constructor. 3. Removed members that were no longer used at all    (in order of disappearance):    (a) fai    (b) testAlpha    (c) nniThreshold    (d) speedUpFromIter 4. Changed the type of the member ref_seq_name (made it const). 5. The Params::Params constructor initializes (I hope!) every    member that needs initializing.  The initializations are    "taken" from parseArgs(). 6. Additional initializations from max_mem_size down (these    members were *not* initialized before).  With comments    explaining what they are. Also... 7. Descoped variables in computePValueChiSquare().
Note: The order in which the members of Params are initialized
doesn't (yet) agree with the order in which they are declared
(in tools.h), so there are still "initialization order" warnings
from the compiler (when -Wreorder-ctor is turned on).
Some day I will reorder both the declaration order in the
header file, and initialization order in the .cpp file, so that
they're easier to read (and, also, so they're consistent).