Fix race condition during model selection
Previously the critical section did not include setting the MF_RUNNING
flag for the selected model. With enough bad luck, a second thread might
enter the critical section before the flag is set and pick up the exact
same model.

Additionally, the modification of current_model, which is shared across threads, was also not protected by the critical section.